Upgrade lib version of playframework from '2.4' to '2.5.1'
Usage of Source instead of Enumerator as recommended in
https://www.playframework.com/documentation/2.5.x/Migration25

Modifications in the SwaggerPlugin constructor:
We do not inject the application, but to avoid a circular dependency
we inject the environment and the configuration.

In order to not modify swagger-scala-module, I've added in the build.sbt
the jackson-module-scala dependence with the same version as the jackson
version used by play 2.5.1.